requirements.txt
python-multipart
google-generativeai
fastapi
streamlit
pandas
pillow
python-dotenv
uvicorn
chromadb
pydantic
uuid
typing
pymilvus
numpy
tqdm
transformers
boto3
openpyxl

torch